+++ app/Http/Controllers/LikeController.php
@@ -11,23 +11,66 @@ class LikeController extends Controller
 {
     public function like(Request $request)
     {
-        $like = new Like;
-        $like->post_id = $request->postId;
-        $like->account_id = $request->accountId;
-        $like->save();
+        $liked = Like::where('post_id', $request->postId)->where('account_id', $request->accountId)->get();
         $post = DB::table('posts')->select('title')->where('id', $request->postId)->get();
-        return response()->json([
-            'You Liked: ' => $post,
-        ]);
+
+        if(count($post) != 0)
+        {
+            if(count($liked) == 0)
+            {
+                $like = new Like;
+                $like->post_id = $request->postId;
+                $like->account_id = $request->accountId;
+                echo $liked;
+                $like->save();
+
+                return response()->json([
+                    'You Liked: ' => $post,
+                ]);
+            }
+            else
+            {
+                return response()->json([
+                    'You liked the post before: ' => $post,
+                ]);
+            }
+        }
+        else
+        {
+          return response()->json([
+              'The post is not available: ' => $post,
+          ]);
+        }
     }
 
     public function unlike(Request $request)
     {
-        Like::where('post_id', $request->postId)->Where('account_id', $request->accountId)->delete();
+        $liked = Like::where('post_id', $request->postId)->where('account_id', $request->accountId)->get();
         $post = DB::table('posts')->select('title')->where('id', $request->postId)->get();
-        return response()->json([
-            'You Unliked: ' => $post,
-        ]);
+
+        if(count($post) != 0)
+        {
+          if(count($liked) != 0)
+          {
+              Like::where('post_id', $request->postId)->Where('account_id', $request->accountId)->delete();
+
+              return response()->json([
+                  'You unliked: ' => $post,
+              ]);
+          }
+          else
+          {
+            return response()->json([
+                'You did not like the post: ' => $post,
+            ]);
+          }
+        }
+        else
+        {
+          return response()->json([
+              'The post is not available: ' => $post,
+          ]);
+        }
     }
 
     public function mostPopular()
@@ -35,7 +78,7 @@ class LikeController extends Controller
         $popularPosts = DB::table('likes')->groupBy('post_id')->having('post_id', '>', 0)->orderBy(DB::raw('count(post_id)'), 'DESC')->get()->toArray();
         $popPostResp = [];
         foreach($popularPosts as $post)
-        { 
+        {
             $popPostResp[] = ['post' => Post::where('id', $post->post_id)->get()];
         }
         return response()->json([
