+++ app/Http/Controllers/PostController.php
@@ -22,35 +22,70 @@ class PostController extends Controller
       ]);
   }
 
-  public function read($title)
+  public function read($id)
   {
-      $post = Post::where('title', $title)->get();
+    $post = Post::where('id', $id)->get();
+
+    if(count($post) != 0)
+    {
       return response()->json([
-          'Account: ' => $post,
+          'Post: ' => $post,
       ]);
+    }
+    else
+    {
+        return response()->json([
+            'The post is not available: ' => $post,
+        ]);
+    }
   }
 
   public function delete(Request $request)
   {
+    $post = DB::table('posts')->select('title')->where('id', $request->id)->whereNull('deleted_at')->get();
+    
+    if(count($post) != 0)
+    {
       Post::where('id', $request->id)->delete();
+      
       return response()->json([
-          'Deleted: ' => $request->id,
+          'Deleted: ' => $post,
       ]);
+    }
+    else
+    {
+        return response()->json([
+            'The post is not available: ' => $post,
+        ]);
+    }
   }
 
   public function update(Request $request)
   {
-      DB::table('posts')
-          ->where('id', $request->id)
-          ->update(['title' => $request->title, 'photo' => $request->photo]);
-      return response()->json([
-          'Title changed to: ' => $request->title,
-      ]);
+    $post = DB::table('posts')->select('title')->where('id', $request->id)->get();
+    
+    if(count($post) != 0)
+    {
+        DB::table('posts')
+            ->where('id', $request->id)
+            ->update(['title' => $request->title, 'photo' => $request->photo]);
+        
+        return response()->json([
+            'Title changed to: ' => $request->title,
+        ]);
+    }
+    else
+    {
+        return response()->json([
+            'The post is not available: ' => $post,
+        ]);
+    }
   }
 
   public function showAllPost($accountId)
   {
       $post = Post::where('account_id', $accountId)->get();
+      
       return response()->json([
           'Account: ' => $post,
       ]);
